requirements.txt
python-binance
numpy
pydispatch
flask
scipy
pytrends